Extract WireDetail component (prep for Item page)
